+++ src/main/java/gaa/tutors/service/TutorService.java
@@ -49,7 +49,7 @@ public class TutorService implements ITutorService {
     }
 
       @Override
-    public void updateTutorById( Long id, String name, String surname, String email, String subject, int cost, float rate) throws ServiceException{
+    public void updateTutorById( Long id, String name, String surname, String email, String subject, Long cost, float rate) throws ServiceException{
         try {
             tutorRepository.updateTutorById(id, name, surname, email, subject, cost, rate);
         } catch (RepositoryException ex) {
